+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Patch
 
 - Internal: Enable React.Strict on documentation (#821)
+- Checkbox/Radiobutton: Updated checkbox and radiobutton borders for disabled state (disabled controls don't show outlines) (#822)
 
 </details>